DJS/requirements.txt
# DJS (Data-based Junction Search) Requirements

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9

# HTTP Client
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Text Processing & NLP
sentence-transformers>=2.2.2
google-genai==1.32.0
openai>=1.3.7
tiktoken>=0.5.2
langchain>=0.0.350
langchain-openai>=0.0.2

# Visualization
networkx>=3.2.1
plotly>=5.17.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Task Scheduling
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
click>=8.1.7
pyyaml>=6.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Logging
loguru>=0.7.2

# Date/Time
python-dateutil>=2.8.2

# Security
bcrypt>=4.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
